Factor each settings pane component out from Settings, factor the common things into lib
