+++ src/views/Settings/TemplateSettings.tsx.temp
@@ -1,0 +1,35 @@
+/*
+ * Copyright (C) 2012-2022  Online-Go.com
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import * as React from "react";
+import { Link } from "react-router-dom";
+import { _, pgettext } from "translate";
+
+import Select from "react-select";
+
+import * as data from "data";
+import { post, get, put, del, getCookie } from "requests";
+import { errorAlerter, errorLogger, ignore } from "misc";
+
+import * as preferences from "preferences";
+import { usePreference } from "preferences";
+
+import { Toggle } from "Toggle";
+
+import { SettingGroupProps, PreferenceToggle, PreferenceLine } from "SettingsCommon";
+
+export
